Fixed files being in the wrong location from #384.
