+++ src/main/resources/assets/mts/lang/en_US.lang
@@ -227,6 +227,9 @@ info.item.engine.bypassratio=Bypass Ratio:
 info.item.engine.maxrpm=Max Possible RPM: 
 info.item.engine.maxsaferpm=Max safe RPM: 
 info.item.engine.fuelconsumption=Fuel use (mb/t): 
+info.item.engine.supercharged=Supercharged
+info.item.engine.superchargerfuelconsumption=Supercharger fuel use (mb/t): 
+info.item.engine.superchargerefficiency=Supercharger efficiency: 
 info.item.engine.fueltype=Fuel type: 
 info.item.engine.hours=Runtime hours: 
 info.item.engine.oilleak=Has an oil leak!
